Initial commit of codeR
codeR: functions that facilitate syntax highlighting of R code in
R plots, in order to facilitate the creation of graphics that
show both the plots and the actual R code. This can be done
using tools such as R-markdown, but I am looking for something
that is more suitable for general graphics rather than reports.

colorise_R.c contains a function that splits a vector of strings
into three lists, containing the line number, the string, and
the class of code. In this initial version the following classes
should be identified:

1. Function calls
2. Quoted text
3. Comments

This highlighting is done without any knowledge of function names
but assuming that function calls will be followed immediately by
an opening '(' character. It seems to work, but there are probably
many edge cases where it will fail, though hopefully without
segmentation errors...

test.R, simply contains some example calls to the function.