Removed recognition of conditionals
Removed the specific recognition of conditional
words as this seemed to be interfering with function
recognition after changing:

while( k > last_end )

to

while( k >= last_end )

The old definition failed for c( ), and presumably for
all single letter functions.

However, I am still getting inconsistent behaviour for the
function recognition with half of the function name not
recognised in many cases.